docs/superpowers/specs/requirements.txt
langchain-openai>=0.2.0
mcp-use>=1.6.0
python-dotenv>=1.0.0
